requirements.txt
# pip 사용자용 (poetry와 동일 버전 권장)
django>=6.0.3,<7
djangorestframework>=3.15,<4
openpyxl>=3.1,<4
pillow>=11,<12
social-auth-app-django>=5.7,<6
python-dotenv>=1.2.2,<2
